Remove duplicated files restored by filter-branch
