+++ app/api/auth/[...nextauth]/route.ts
@@ -1,6 +1,6 @@
 import { UserI } from "@/interfaces/users";
 import { verifyPassword } from "@/lib/auth/password";
-import { connectToDB } from "@/lib/db";
+import { connectToDB, disconnectFromDB } from "@/lib/db";
 import NextAuth from "next-auth"
 import CredentialsProvider from "next-auth/providers/credentials";
 
@@ -24,12 +24,14 @@ export const authOptions = {
 				const usersCollection = db.collection('users');
 				const user = await usersCollection.findOne({ email: credentials.email }) as UserI | null;
 				if (!user) {
-					client.close();
+					// client.close();
+					await disconnectFromDB();
 					throw new Error("No user found");
 				}
 				const isValid = await verifyPassword(password, user.password || '');
 				if (!isValid) { client.close(); throw new Error('Email or password invalid'); }
-				client.close();
+				// client.close();
+				await disconnectFromDB();
 				return {
 					id: user._id?.toString() || user.id || '',
 					email: user.email,
